+++ attributes/default.rb
@@ -127,6 +127,24 @@ default['openstack']['block-storage']['nfs']['mount_point_base'] = '/mnt/cinder-
 default['openstack']['block-storage']['nfs']['nfs_disk_util'] = 'df'
 default['openstack']['block-storage']['nfs']['nfs_sparsed_volumes'] = 'true'
 
+# Storwize/SVC Support
+default['openstack']['block-storage']['san']['san_ip'] = '127.0.0.1'
+default['openstack']['block-storage']['san']['san_login'] = 'admin'
+default['openstack']['block-storage']['san']['san_private_key'] = '/v7000_rsa'
+default['openstack']['block-storage']['storwize']['storwize_svc_volpool_name'] = 'volpool'
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_rsize'] = '2%'
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_warning'] = 0
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_autoexpand'] = true
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_grainsize'] = 256
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_compression'] = false
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_easytier'] = true
+default['openstack']['block-storage']['storwize']['storwize_svc_flashcopy_timeout'] = 120
+default['openstack']['block-storage']['storwize']['storwize_svc_vol_iogrp'] = 0
+default['openstack']['block-storage']['storwize']['storwize_svc_connection_protocol'] = 'iSCSI'
+default['openstack']['block-storage']['storwize']['storwize_svc_iscsi_chap_enabled'] = true
+default['openstack']['block-storage']['storwize']['storwize_svc_multipath_enabled'] = false
+default['openstack']['block-storage']['storwize']['storwize_svc_multihostmap_enabled'] = true
+
 # logging attribute
 default['openstack']['block-storage']['syslog']['use'] = false
 default['openstack']['block-storage']['syslog']['facility'] = 'LOG_LOCAL2'